gitignore 적용
