+++ mainSite/objectRecog/templates/objectrecogMain.html
@@ -1,10 +1,11 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
+  <head>
+    <meta charset="UTF-8" />
     <title>Title</title>
-</head>
-<body>
-물체인식 초기 페이지입니다.
-</body>
+  </head>
+  <body>
+    물체인식 초기 페이지입니다.
+    <a href="{% url 'kakaoApi' %}">kakaoApi</a>
+  </body>
 </html>
