+++ mainSite/objectRecog/urls.py
@@ -3,5 +3,6 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path('', views.objectIndex, name='objectIndex'),
+  path('', views.objectIndex, name='objectIndex'),
+  path('kakaoApi/', views.kakaoApi, name="kakaoApi"),
 ]
